requirements.txt
Django==4.0.5
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
httplib2==0.20.4
openai==0.27.8
pexpect==4.8.0
psycopg2==2.9.3
psycopg2-binary==2.9.3